init branch with not recommended workaround

+++ src/ExperimentDataWrangler.cpp
@@ -207,12 +207,12 @@ ExperimentData* ExperimentDataWrangler::loadUCI(const UCINames &dataset) {
     UCI raw_uci(dataset, path_prefix);
     DataSet *uci_training = new DataSet {
         FeatureVectorHeader(raw_uci.getTrainingData()[0].x.size(), FeatureType::NUMERIC), // XXX many assumptions
-        raw_uci.getLabels().size(),
+        (int)(raw_uci.getLabels().size()),
         std::vector<DataRow>(raw_uci.getTrainingData().size())
     };
     DataSet *uci_test = new DataSet {
         FeatureVectorHeader(raw_uci.getTestData()[0].x.size(), FeatureType::NUMERIC),
-        raw_uci.getLabels().size(),
+        (int)(raw_uci.getLabels().size()),
         std::vector<DataRow>(raw_uci.getTestData().size())
     };
